chore: remove service
