+++ apps/taiga-lumbermill/src/dashboards/dashboards-list/dashboards-list.component.html
@@ -1,4 +1,4 @@
-@for (card of dashboardsService; track $index) {
+@for (card of dashboardsListData; track $index) {
     <a
         tuiCardLarge
         tuiHeader
